+++ hellorenaud/src/txt_scroll_hard.h
@@ -1,6 +1,6 @@
 #ifndef _TXT_SCROLL_H_
 #define _TXT_SCROLL_H_
 
-void scroll_hard(unsigned int step, u8* screen_plot_address);
+void scroll_hard(u16 step, u8* screen_plot_address);
 
 #endif
